locked token pos creator - part 1
