+++ auto-pos-creator/src/multi_contract_interactions/mod.rs
@@ -1,2 +1,4 @@
 pub mod create_pos;
+pub mod create_pos_endpoints;
 pub mod exit_pos;
+pub mod exit_pos_endpoints;
